--- Assets/A-Detonation.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: fee58e489ea43a6488fd5a504fafe8d1
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
