--- Assets/Basement/Restroom/mirror.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: beb3095fa7510934d9138f9716b10065
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
