--- Assets/Resources/Items.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: cb81267bdd3ec2c4cbf2286c1381c70f
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
