+++ Assets/Scripts/IngameCtrl.cs
@@ -1,0 +1,55 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class IngameCtrl : MonoBehaviour
+{
+  //  [Header("TextTriggerScript")]
+    //public TextTrigger m_TextTrigger;
+    public TextTrigger m_Room1_Board_Text;
+
+    [Header("TextTriggerUI")]
+    public GameObject m_Room1_Board_UI;
+
+
+    [Header("KeyBoardBool")]
+    public bool m_pressR;
+    public bool m_pressZ;
+
+
+    // Update is called once per frame
+    void Update()
+    {
+        if (Input.GetKeyDown(KeyCode.R))
+        {
+            m_pressR = true;
+            Debug.Log("press R");
+        }
+
+        if (Input.GetKeyUp(KeyCode.R))
+        {
+            m_pressR = false;
+            Debug.Log("R에서 손뗌");
+        }
+
+        if (Input.GetKeyDown(KeyCode.Z))
+        {
+            m_pressZ = true;
+            Debug.Log("press Z");
+        }
+        if (Input.GetKeyUp(KeyCode.Z))
+        {
+            m_pressZ = false;
+            Debug.Log("z에서 손 뗌");
+        }
+
+       if(m_pressR && m_Room1_Board_Text.b_textTrigger)
+        {
+            m_Room1_Board_UI.gameObject.SetActive(true);
+        }
+        else
+        {
+            m_Room1_Board_UI.gameObject.SetActive(false);
+        }
+    }
+}
