+++ Assets/Scripts/IngameCtrl.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7692893ca3882fa40920b6232855a62d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
