--- Assets/legacy/Assets/AssetArchieve/Classic Picture Frame.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6c7083ba8b1734ca4aa9c9cbeba3fa52
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
