--- Assets/legacy/Assets/AssetArchieve/Conifers [BOTD]/CTI Runtime Components/Default Components.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e39383e61349b4445966d2747d1798c7
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
