--- Assets/legacy/Assets/AssetArchieve/Conifers [BOTD]/Sources/Billboards.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: eed1d7eed3f9845cf8db87a0770ca745
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
