--- Assets/legacy/Assets/AssetArchieve/Screaming statue.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f96ced33ca4d74a5d9bd1569e0b9284d
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
